updated artefacts

+++ Assets/Scripts/Singletons/InventoryManager.cs
@@ -13,7 +13,7 @@ namespace GM.Inventory
     {
         public static InventoryManager Instance = null;
 
-        public int prestigePoints;
+        public long prestigePoints;
 
         public int BlueGems;
 
@@ -39,7 +39,7 @@ namespace GM.Inventory
             BountyPoints = node.GetValueOrDefault("bountyPoints", 0).AsInt;
             ArmouryPoints = node.GetValueOrDefault("armouryPoints", 0).AsInt;
 
-            prestigePoints = node.GetValueOrDefault("prestigePoints", 0).AsInt;
+            prestigePoints = node.GetValueOrDefault("prestigePoints", 0).AsLong;
         }
     }
 }
